Decoupling: Move WP menu handling to wp-dashboard
